% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Annotation}
\alias{Annotation}
\title{Genomics API Objects 
Provides access to Genomics data.}
\usage{
Annotation(Annotation.info = NULL, annotationSetId = NULL, id = NULL,
  info = NULL, name = NULL, position = NULL, transcript = NULL,
  type = NULL, variant = NULL)
}
\arguments{
\item{Annotation.info}{The \link{Annotation.info} object or list of objects}

\item{annotationSetId}{The annotation set to which this annotation belongs}

\item{id}{The server-generated annotation ID, unique across all annotations}

\item{info}{A string which maps to an array of values}

\item{name}{The display name of this annotation}

\item{position}{The position of this annotation on the reference sequence}

\item{transcript}{A transcript value represents the assertion that a particular region of the reference genome may be transcribed as RNA}

\item{type}{The data type for this annotation}

\item{variant}{A variant annotation, which describes the effect of a variant on the genome, the coding sequence, and/or higher level consequences at the organism level e}
}
\value{
Annotation object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:14:37
filename: /Users/mark/dev/R/autoGoogleAPI/googlegenomicsv1beta2.auto/R/genomics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Annotation Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An annotation describes a region of reference genome. The value of an annotation may be one of several canonical types, supplemented by arbitrary info tags. An annotation is not inherently associated with a specific sample or individual (though a client could choose to use annotations in this way). Example canonical annotation types are GENE and VARIANT.
}
\seealso{
Other Annotation functions: \code{\link{Annotation.info}},
  \code{\link{annotations.create}},
  \code{\link{annotations.patch}},
  \code{\link{annotations.update}}
}

