% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{AnnotationSet.info}
\alias{AnnotationSet.info}
\title{AnnotationSet.info Object}
\usage{
AnnotationSet.info()
}
\value{
AnnotationSet.info object
}
\description{
AnnotationSet.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A string which maps to an array of values.
}
\seealso{
Other AnnotationSet functions: \code{\link{AnnotationSet}},
  \code{\link{annotationSets.create}},
  \code{\link{annotationSets.patch}},
  \code{\link{annotationSets.update}}
}

