% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{annotationSets.create}
\alias{annotationSets.create}
\title{Creates a new annotation set. Caller must have WRITE permission for the associated dataset.The following fields must be provided when creating an annotation set:  - datasetId - referenceSetId  All other fields may be optionally specified, unless documented as being server-generated (for example, the id field).}
\usage{
annotationSets.create(AnnotationSet)
}
\arguments{
\item{AnnotationSet}{The \link{AnnotationSet} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}

Other AnnotationSet functions: \code{\link{AnnotationSet.info}},
  \code{\link{AnnotationSet}},
  \code{\link{annotationSets.patch}},
  \code{\link{annotationSets.update}}
}

