% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{BatchAnnotationsResponse}
\alias{BatchAnnotationsResponse}
\title{BatchAnnotationsResponse Object}
\usage{
BatchAnnotationsResponse(entries = NULL)
}
\arguments{
\item{entries}{The resulting per-annotation entries, ordered consistently with the original request}
}
\value{
BatchAnnotationsResponse object
}
\description{
BatchAnnotationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

