% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{BatchAnnotationsResponseEntry}
\alias{BatchAnnotationsResponseEntry}
\title{BatchAnnotationsResponseEntry Object}
\usage{
BatchAnnotationsResponseEntry(annotation = NULL, status = NULL)
}
\arguments{
\item{annotation}{The annotation, if any}

\item{status}{The resulting status for this annotation operation}
}
\value{
BatchAnnotationsResponseEntry object
}
\description{
BatchAnnotationsResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

