% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{BatchAnnotationsResponseEntryStatus}
\alias{BatchAnnotationsResponseEntryStatus}
\title{BatchAnnotationsResponseEntryStatus Object}
\usage{
BatchAnnotationsResponseEntryStatus(code = NULL, message = NULL)
}
\arguments{
\item{code}{The HTTP status code for this operation}

\item{message}{Error message for this status, if any}
}
\value{
BatchAnnotationsResponseEntryStatus object
}
\description{
BatchAnnotationsResponseEntryStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}

