% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Call.info}
\alias{Call.info}
\title{Call.info Object}
\usage{
Call.info()
}
\value{
Call.info object
}
\description{
Call.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A string which maps to an array of values.
}
\seealso{
Other Call functions: \code{\link{Call}}
}

