% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Call}
\alias{Call}
\title{Call Object}
\usage{
Call(Call.info = NULL, callSetId = NULL, callSetName = NULL,
  genotype = NULL, genotypeLikelihood = NULL, info = NULL,
  phaseset = NULL)
}
\arguments{
\item{Call.info}{The \link{Call.info} object or list of objects}

\item{callSetId}{The ID of the call set this variant call belongs to}

\item{callSetName}{The name of the call set this variant call belongs to}

\item{genotype}{The genotype of this variant call}

\item{genotypeLikelihood}{The genotype likelihoods for this variant call}

\item{info}{A string which maps to an array of values}

\item{phaseset}{If this field is present, this variant call's genotype ordering implies the phase of the bases and is consistent with any other variant calls in the same reference sequence which have the same phaseset value}
}
\value{
Call object
}
\description{
Call Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A call represents the determination of genotype with respect to a particular variant. It may include associated information such as quality and phasing. For example, a call might assign a probability of 0.32 to the occurrence of a SNP named rs1234 in a call set with the name NA12345.
}
\seealso{
Other Call functions: \code{\link{Call.info}}
}

