% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{CallSet}
\alias{CallSet}
\title{CallSet Object}
\usage{
CallSet(CallSet.info = NULL, created = NULL, id = NULL, info = NULL,
  name = NULL, sampleId = NULL, variantSetIds = NULL)
}
\arguments{
\item{CallSet.info}{The \link{CallSet.info} object or list of objects}

\item{created}{The date this call set was created in milliseconds from the epoch}

\item{id}{The Google generated ID of the call set, immutable}

\item{info}{A string which maps to an array of values}

\item{name}{The call set name}

\item{sampleId}{The sample ID this call set corresponds to}

\item{variantSetIds}{The IDs of the variant sets this call set belongs to}
}
\value{
CallSet object
}
\description{
CallSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A call set is a collection of variant calls, typically for one sample. It belongs to a variant set.
}
\seealso{
Other CallSet functions: \code{\link{CallSet.info}},
  \code{\link{callsets.create}},
  \code{\link{callsets.patch}},
  \code{\link{callsets.update}}
}

