% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Object}
\usage{
Dataset(createTime = NULL, id = NULL, isPublic = NULL, name = NULL,
  projectNumber = NULL)
}
\arguments{
\item{createTime}{The time this dataset was created, in seconds from the epoch}

\item{id}{The Google generated ID of the dataset, immutable}

\item{isPublic}{Flag indicating whether or not a dataset is publicly viewable}

\item{name}{The dataset name}

\item{projectNumber}{The Google Developers Console project number that this dataset belongs to}
}
\value{
Dataset object
}
\description{
Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Dataset is a collection of genomic data.
}
\seealso{
Other Dataset functions: \code{\link{datasets.create}},
  \code{\link{datasets.patch}},
  \code{\link{datasets.update}}
}

