% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{experimental.jobs.create}
\alias{experimental.jobs.create}
\title{Creates and asynchronously runs an ad-hoc job. This is an experimental call and may be removed or changed at any time.}
\usage{
experimental.jobs.create(ExperimentalCreateJobRequest)
}
\arguments{
\item{ExperimentalCreateJobRequest}{The \link{ExperimentalCreateJobRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.read_write
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.read_write, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}

Other ExperimentalCreateJobRequest functions: \code{\link{ExperimentalCreateJobRequest}}
}

