% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExperimentalCreateJobRequest}
\alias{ExperimentalCreateJobRequest}
\title{ExperimentalCreateJobRequest Object}
\usage{
ExperimentalCreateJobRequest(align = NULL, callVariants = NULL,
  gcsOutputPath = NULL, pairedSourceUris = NULL, projectNumber = NULL,
  sourceUris = NULL)
}
\arguments{
\item{align}{Specifies whether or not to run the alignment pipeline}

\item{callVariants}{Specifies whether or not to run the variant calling pipeline}

\item{gcsOutputPath}{Specifies where to copy the results of certain pipelines}

\item{pairedSourceUris}{A list of Google Cloud Storage URIs of paired end}

\item{projectNumber}{Required}

\item{sourceUris}{A list of Google Cloud Storage URIs of data files to operate upon}
}
\value{
ExperimentalCreateJobRequest object
}
\description{
ExperimentalCreateJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The job creation request.
}
\seealso{
Other ExperimentalCreateJobRequest functions: \code{\link{experimental.jobs.create}}
}

