% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExperimentalCreateJobResponse}
\alias{ExperimentalCreateJobResponse}
\title{ExperimentalCreateJobResponse Object}
\usage{
ExperimentalCreateJobResponse(jobId = NULL)
}
\arguments{
\item{jobId}{A job ID that can be used to get status information}
}
\value{
ExperimentalCreateJobResponse object
}
\description{
ExperimentalCreateJobResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The job creation response.
}

