% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExportReadGroupSetsRequest}
\alias{ExportReadGroupSetsRequest}
\title{ExportReadGroupSetsRequest Object}
\usage{
ExportReadGroupSetsRequest(exportUri = NULL, projectNumber = NULL,
  readGroupSetIds = NULL, referenceNames = NULL)
}
\arguments{
\item{exportUri}{Required}

\item{projectNumber}{Required}

\item{readGroupSetIds}{Required}

\item{referenceNames}{The reference names to export}
}
\value{
ExportReadGroupSetsRequest object
}
\description{
ExportReadGroupSetsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read group set export request.
}
\seealso{
Other ExportReadGroupSetsRequest functions: \code{\link{readgroupsets.export}}
}

