% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExportReadGroupSetsResponse}
\alias{ExportReadGroupSetsResponse}
\title{ExportReadGroupSetsResponse Object}
\usage{
ExportReadGroupSetsResponse(jobId = NULL)
}
\arguments{
\item{jobId}{A job ID that can be used to get status information}
}
\value{
ExportReadGroupSetsResponse object
}
\description{
ExportReadGroupSetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read group set export response.
}

