% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExportVariantSetRequest}
\alias{ExportVariantSetRequest}
\title{ExportVariantSetRequest Object}
\usage{
ExportVariantSetRequest(bigqueryDataset = NULL, bigqueryTable = NULL,
  callSetIds = NULL, format = NULL, projectNumber = NULL)
}
\arguments{
\item{bigqueryDataset}{Required}

\item{bigqueryTable}{Required}

\item{callSetIds}{If provided, only variant call information from the specified call sets will be exported}

\item{format}{The format for the exported data}

\item{projectNumber}{Required}
}
\value{
ExportVariantSetRequest object
}
\description{
ExportVariantSetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data export request.
}
\seealso{
Other ExportVariantSetRequest functions: \code{\link{variantsets.export}}
}

