% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExportVariantSetResponse}
\alias{ExportVariantSetResponse}
\title{ExportVariantSetResponse Object}
\usage{
ExportVariantSetResponse(jobId = NULL)
}
\arguments{
\item{jobId}{A job ID that can be used to get status information}
}
\value{
ExportVariantSetResponse object
}
\description{
ExportVariantSetResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data export response.
}

