% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ExternalId}
\alias{ExternalId}
\title{ExternalId Object}
\usage{
ExternalId(id = NULL, sourceName = NULL)
}
\arguments{
\item{id}{The id used by the source of this data}

\item{sourceName}{The name of the source of this data}
}
\value{
ExternalId object
}
\description{
ExternalId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

