% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ImportReadGroupSetsResponse}
\alias{ImportReadGroupSetsResponse}
\title{ImportReadGroupSetsResponse Object}
\usage{
ImportReadGroupSetsResponse(jobId = NULL)
}
\arguments{
\item{jobId}{A job ID that can be used to get status information}
}
\value{
ImportReadGroupSetsResponse object
}
\description{
ImportReadGroupSetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read group set import response.
}

