% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ImportVariantsRequest}
\alias{ImportVariantsRequest}
\title{ImportVariantsRequest Object}
\usage{
ImportVariantsRequest(format = NULL, normalizeReferenceNames = NULL,
  sourceUris = NULL)
}
\arguments{
\item{format}{The format of the variant data being imported}

\item{normalizeReferenceNames}{Convert reference names to the canonical representation}

\item{sourceUris}{A list of URIs referencing variant files in Google Cloud Storage}
}
\value{
ImportVariantsRequest object
}
\description{
ImportVariantsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data import request.
}
\seealso{
Other ImportVariantsRequest functions: \code{\link{variantsets.importVariants}}
}

