% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ImportVariantsResponse}
\alias{ImportVariantsResponse}
\title{ImportVariantsResponse Object}
\usage{
ImportVariantsResponse(jobId = NULL)
}
\arguments{
\item{jobId}{A job ID that can be used to get status information}
}
\value{
ImportVariantsResponse object
}
\description{
ImportVariantsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant data import response.
}

