% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Int32Value}
\alias{Int32Value}
\title{Int32Value Object}
\usage{
Int32Value(value = NULL)
}
\arguments{
\item{value}{The int32 value}
}
\value{
Int32Value object
}
\description{
Int32Value Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Wrapper message for `int32`.The JSON representation for `Int32Value` is JSON number.
}

