% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Job}
\alias{Job}
\title{Job Object}
\usage{
Job(created = NULL, detailedStatus = NULL, errors = NULL, id = NULL,
  importedIds = NULL, projectNumber = NULL, request = NULL,
  status = NULL, warnings = NULL)
}
\arguments{
\item{created}{The date this job was created, in milliseconds from the epoch}

\item{detailedStatus}{A more detailed description of this job's current status}

\item{errors}{Any errors that occurred during processing}

\item{id}{The job ID}

\item{importedIds}{If this Job represents an import, this field will contain the IDs of the objects that were successfully imported}

\item{projectNumber}{The Google Developers Console project number to which this job belongs}

\item{request}{A summarized representation of the original service request}

\item{status}{The status of this job}

\item{warnings}{Any warnings that occurred during processing}
}
\value{
Job object
}
\description{
Job Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Job represents an ongoing process that can be monitored for status information.
}

