% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{JobRequest}
\alias{JobRequest}
\title{JobRequest Object}
\usage{
JobRequest(destination = NULL, source = NULL, type = NULL)
}
\arguments{
\item{destination}{The data destination of the request, for example, a Google BigQuery Table or Dataset ID}

\item{source}{The data source of the request, for example, a Google Cloud Storage object path or Readset ID}

\item{type}{The original request type}
}
\value{
JobRequest object
}
\description{
JobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A summary representation of the service request that spawned the job.
}

