% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{jobs.cancel}
\alias{jobs.cancel}
\title{Cancels a job by ID. Note that it is possible for partial results to be generated and stored for cancelled jobs.}
\usage{
jobs.cancel(jobId)
}
\arguments{
\item{jobId}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}
}

