% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{KeyValue}
\alias{KeyValue}
\title{KeyValue Object}
\usage{
KeyValue(key = NULL, value = NULL)
}
\arguments{
\item{key}{A string which maps to an array of values}

\item{value}{The string values}
}
\value{
KeyValue object
}
\description{
KeyValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Used to hold basic key value information.
}

