% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ListCoverageBucketsResponse}
\alias{ListCoverageBucketsResponse}
\title{ListCoverageBucketsResponse Object}
\usage{
ListCoverageBucketsResponse(bucketWidth = NULL, coverageBuckets = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{bucketWidth}{The length of each coverage bucket in base pairs}

\item{coverageBuckets}{The coverage buckets}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}
}
\value{
ListCoverageBucketsResponse object
}
\description{
ListCoverageBucketsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

