% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{MergeVariantsRequest}
\alias{MergeVariantsRequest}
\title{MergeVariantsRequest Object}
\usage{
MergeVariantsRequest(variants = NULL)
}
\arguments{
\item{variants}{The variants to be merged with existing variants}
}
\value{
MergeVariantsRequest object
}
\description{
MergeVariantsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other MergeVariantsRequest functions: \code{\link{variantsets.mergeVariants}}
}

