% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Metadata.info}
\alias{Metadata.info}
\title{Metadata.info Object}
\usage{
Metadata.info()
}
\value{
Metadata.info object
}
\description{
Metadata.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A string which maps to an array of values.
}
\seealso{
Other Metadata functions: \code{\link{Metadata}}
}

