% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Metadata}
\alias{Metadata}
\title{Metadata Object}
\usage{
Metadata(Metadata.info = NULL, description = NULL, id = NULL,
  info = NULL, key = NULL, number = NULL, type = NULL, value = NULL)
}
\arguments{
\item{Metadata.info}{The \link{Metadata.info} object or list of objects}

\item{description}{A textual description of this metadata}

\item{id}{User-provided ID field, not enforced by this API}

\item{info}{A string which maps to an array of values}

\item{key}{The top-level key}

\item{number}{The number of values that can be included in a field described by this metadata}

\item{type}{The type of data}

\item{value}{The value field for simple metadata}
}
\value{
Metadata object
}
\description{
Metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describes a single piece of variant call metadata. These data include a top level key and either a single value string (value) or a list of key-value pairs (info.) Value and info are mutually exclusive.
}
\seealso{
Other Metadata functions: \code{\link{Metadata.info}}
}

