% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{QueryRange}
\alias{QueryRange}
\title{QueryRange Object}
\usage{
QueryRange(end = NULL, referenceId = NULL, referenceName = NULL,
  start = NULL)
}
\arguments{
\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{referenceId}{The ID of the reference to query}

\item{referenceName}{The name of the reference to query, within the reference set associated with this query}

\item{start}{The start position of the range on the reference, 0-based inclusive}
}
\value{
QueryRange object
}
\description{
QueryRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A 0-based half-open genomic coordinate range for search requests. reference_id or reference_name must be set.
}

