% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Range}
\alias{Range}
\title{Range Object}
\usage{
Range(end = NULL, referenceName = NULL, start = NULL)
}
\arguments{
\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{referenceName}{The reference sequence name, for example chr1, 1, or chrX}

\item{start}{The start position of the range on the reference, 0-based inclusive}
}
\value{
Range object
}
\description{
Range Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A 0-based half-open genomic coordinate range over a reference sequence.
}

