% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{RangePosition}
\alias{RangePosition}
\title{RangePosition Object}
\usage{
RangePosition(end = NULL, referenceId = NULL, referenceName = NULL,
  reverseStrand = NULL, start = NULL)
}
\arguments{
\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{referenceId}{The ID of the Google Genomics reference associated with this range}

\item{referenceName}{The display name corresponding to the reference specified by referenceId, for example chr1, 1, or chrX}

\item{reverseStrand}{Whether this range refers to the reverse strand, as opposed to the forward strand}

\item{start}{The start position of the range on the reference, 0-based inclusive}
}
\value{
RangePosition object
}
\description{
RangePosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A 0-based half-open genomic coordinate range over a reference sequence, for representing the position of a genomic resource.
}

