% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Read}
\alias{Read}
\title{Read Object}
\usage{
Read(Read.info = NULL, alignedQuality = NULL, alignedSequence = NULL,
  alignment = NULL, duplicateFragment = NULL,
  failedVendorQualityChecks = NULL, fragmentLength = NULL,
  fragmentName = NULL, id = NULL, info = NULL, nextMatePosition = NULL,
  numberReads = NULL, properPlacement = NULL, readGroupId = NULL,
  readGroupSetId = NULL, readNumber = NULL, secondaryAlignment = NULL,
  supplementaryAlignment = NULL)
}
\arguments{
\item{Read.info}{The \link{Read.info} object or list of objects}

\item{alignedQuality}{The quality of the read sequence contained in this alignment record}

\item{alignedSequence}{The bases of the read sequence contained in this alignment record, without CIGAR operations applied (equivalent to SEQ in SAM)}

\item{alignment}{The linear alignment for this alignment record}

\item{duplicateFragment}{The fragment is a PCR or optical duplicate (SAM flag 0x400)}

\item{failedVendorQualityChecks}{Whether this read did not pass filters, such as platform or vendor quality controls (SAM flag 0x200)}

\item{fragmentLength}{The observed length of the fragment, equivalent to TLEN in SAM}

\item{fragmentName}{The fragment name}

\item{id}{The unique ID for this read}

\item{info}{A string which maps to an array of values}

\item{nextMatePosition}{The position of the primary alignment of the (readNumber+1)%numberReads read in the fragment}

\item{numberReads}{The number of reads in the fragment (extension to SAM flag 0x1)}

\item{properPlacement}{The orientation and the distance between reads from the fragment are consistent with the sequencing protocol (SAM flag 0x2)}

\item{readGroupId}{The ID of the read group this read belongs to}

\item{readGroupSetId}{The ID of the read group set this read belongs to}

\item{readNumber}{The read number in sequencing}

\item{secondaryAlignment}{Whether this alignment is secondary}

\item{supplementaryAlignment}{Whether this alignment is supplementary}
}
\value{
Read object
}
\description{
Read Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A read alignment describes a linear alignment of a string of DNA to a reference sequence, in addition to metadata about the fragment (the molecule of DNA sequenced) and the read (the bases which were read by the sequencer). A read is equivalent to a line in a SAM file. A read belongs to exactly one read group and exactly one read group set. Generating a reference-aligned sequence string When interacting with mapped reads, it's often useful to produce a string representing the local alignment of the read to reference. The following pseudocode demonstrates one way of doing this:out = '' offset = 0 for c in read.alignment.cigar { switch c.operation { case 'ALIGNMENT_MATCH', 'SEQUENCE_MATCH', 'SEQUENCE_MISMATCH': out += read.alignedSequence[offset:offset+c.operationLength] offset += c.operationLength break case 'CLIP_SOFT', 'INSERT': offset += c.operationLength break case 'PAD': out += repeat('*', c.operationLength) break case 'DELETE': out += repeat('-', c.operationLength) break case 'SKIP': out += repeat(' ', c.operationLength) break case 'CLIP_HARD': break } } return outConverting to SAM's CIGAR string The following pseudocode generates a SAM CIGAR string from the cigar field. Note that this is a lossy conversion (cigar.referenceSequence is lost).cigarMap = { 'ALIGNMENT_MATCH': 'M', 'INSERT': 'I', 'DELETE': 'D', 'SKIP': 'N', 'CLIP_SOFT': 'S', 'CLIP_HARD': 'H', 'PAD': 'P', 'SEQUENCE_MATCH': '=', 'SEQUENCE_MISMATCH': 'X', } cigarStr = '' for c in read.alignment.cigar { cigarStr += c.operationLength + cigarMap[c.operation] } return cigarStr
}
\seealso{
Other Read functions: \code{\link{Read.info}}
}

