% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroup}
\alias{ReadGroup}
\title{ReadGroup Object}
\usage{
ReadGroup(ReadGroup.info = NULL, datasetId = NULL, description = NULL,
  experiment = NULL, id = NULL, info = NULL, name = NULL,
  predictedInsertSize = NULL, programs = NULL, referenceSetId = NULL,
  sampleId = NULL)
}
\arguments{
\item{ReadGroup.info}{The \link{ReadGroup.info} object or list of objects}

\item{datasetId}{The ID of the dataset this read group belongs to}

\item{description}{A free-form text description of this read group}

\item{experiment}{The experiment used to generate this read group}

\item{id}{The generated unique read group ID}

\item{info}{A string which maps to an array of values}

\item{name}{The read group name}

\item{predictedInsertSize}{The predicted insert size of this read group}

\item{programs}{The programs used to generate this read group}

\item{referenceSetId}{The reference set to which the reads in this read group are aligned}

\item{sampleId}{The sample this read group's data was generated from}
}
\value{
ReadGroup object
}
\description{
ReadGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A read group is all the data that's processed the same way by the sequencer.
}
\seealso{
Other ReadGroup functions: \code{\link{ReadGroup.info}}
}

