% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroupExperiment}
\alias{ReadGroupExperiment}
\title{ReadGroupExperiment Object}
\usage{
ReadGroupExperiment(instrumentModel = NULL, libraryId = NULL,
  platformUnit = NULL, sequencingCenter = NULL)
}
\arguments{
\item{instrumentModel}{The instrument model used as part of this experiment}

\item{libraryId}{The library used as part of this experiment}

\item{platformUnit}{The platform unit used as part of this experiment e}

\item{sequencingCenter}{The sequencing center used as part of this experiment}
}
\value{
ReadGroupExperiment object
}
\description{
ReadGroupExperiment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

