% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroupProgram}
\alias{ReadGroupProgram}
\title{ReadGroupProgram Object}
\usage{
ReadGroupProgram(commandLine = NULL, id = NULL, name = NULL,
  prevProgramId = NULL, version = NULL)
}
\arguments{
\item{commandLine}{The command line used to run this program}

\item{id}{The user specified locally unique ID of the program}

\item{name}{The display name of the program}

\item{prevProgramId}{The ID of the program run before this one}

\item{version}{The version of the program run}
}
\value{
ReadGroupProgram object
}
\description{
ReadGroupProgram Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

