% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReadGroupSet}
\alias{ReadGroupSet}
\title{ReadGroupSet Object}
\usage{
ReadGroupSet(ReadGroupSet.info = NULL, datasetId = NULL, filename = NULL,
  id = NULL, info = NULL, name = NULL, readGroups = NULL,
  referenceSetId = NULL)
}
\arguments{
\item{ReadGroupSet.info}{The \link{ReadGroupSet.info} object or list of objects}

\item{datasetId}{The dataset ID}

\item{filename}{The filename of the original source file for this read group set, if any}

\item{id}{The read group set ID}

\item{info}{A string which maps to an array of values}

\item{name}{The read group set name}

\item{readGroups}{The read groups in this set}

\item{referenceSetId}{The reference set the reads in this read group set are aligned to}
}
\value{
ReadGroupSet object
}
\description{
ReadGroupSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A read group set is a logical collection of read groups, which are collections of reads produced by a sequencer. A read group set typically models reads corresponding to one sample, sequenced one way, and aligned one way.  - A read group set belongs to one dataset.- A read group belongs to one read group set.- A read belongs to one read group.
}
\seealso{
Other ReadGroupSet functions: \code{\link{ReadGroupSet.info}},
  \code{\link{readgroupsets.patch}},
  \code{\link{readgroupsets.update}}
}

