% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{readgroupsets.import}
\alias{readgroupsets.import}
\title{Creates read group sets by asynchronously importing the provided information. The caller must have WRITE permissions to the dataset.Notes on BAM import:  - Tags will be converted to strings - tag types are not preserved- Comments (@CO) in the input file header are not imported- Original order of reference headers is not preserved- Any reverse stranded unmapped reads will be reverse complemented, and their qualities (also the 'BQ' and 'OQ' tags, if any) will be reversed- Unmapped reads will be stripped of positional information (referenceName and position)}
\usage{
readgroupsets.import(ImportReadGroupSetsRequest)
}
\arguments{
\item{ImportReadGroupSetsRequest}{The \link{ImportReadGroupSetsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.read_write
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.read_write, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}

Other ImportReadGroupSetsRequest functions: \code{\link{ImportReadGroupSetsRequest}}
}

