% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Reference}
\alias{Reference}
\title{Reference Object}
\usage{
Reference(id = NULL, length = NULL, md5checksum = NULL, name = NULL,
  ncbiTaxonId = NULL, sourceAccessions = NULL, sourceURI = NULL)
}
\arguments{
\item{id}{The Google generated immutable ID of the reference}

\item{length}{The length of this reference's sequence}

\item{md5checksum}{MD5 of the upper-case sequence excluding all whitespace characters (this is equivalent to SQ:M5 in SAM)}

\item{name}{The name of this reference, for example 22}

\item{ncbiTaxonId}{ID from http://www}

\item{sourceAccessions}{All known corresponding accession IDs in INSDC (GenBank/ENA/DDBJ) ideally with a version number, for example GCF_000001405}

\item{sourceURI}{The URI from which the sequence was obtained}
}
\value{
Reference object
}
\description{
Reference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference is a canonical assembled DNA sequence, intended to act as a reference coordinate space for other genomic annotations. A single reference might represent the human chromosome 1 or mitochandrial DNA, for instance. A reference belongs to one or more reference sets.
}

