% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReferenceBound}
\alias{ReferenceBound}
\title{ReferenceBound Object}
\usage{
ReferenceBound(referenceName = NULL, upperBound = NULL)
}
\arguments{
\item{referenceName}{The name of the reference associated with this reference bound}

\item{upperBound}{An upper bound (inclusive) on the starting coordinate of any variant in the reference sequence}
}
\value{
ReferenceBound object
}
\description{
ReferenceBound Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ReferenceBound records an upper bound for the starting coordinate of variants in a particular reference.
}

