% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{ReferenceSet}
\alias{ReferenceSet}
\title{ReferenceSet Object}
\usage{
ReferenceSet(assemblyId = NULL, description = NULL, id = NULL,
  md5checksum = NULL, ncbiTaxonId = NULL, referenceIds = NULL,
  sourceAccessions = NULL, sourceURI = NULL)
}
\arguments{
\item{assemblyId}{Public id of this reference set, such as GRCh37}

\item{description}{Free text description of this reference set}

\item{id}{The Google generated immutable ID of the reference set}

\item{md5checksum}{Order-independent MD5 checksum which identifies this reference set}

\item{ncbiTaxonId}{ID from http://www}

\item{referenceIds}{The IDs of the reference objects that are part of this set}

\item{sourceAccessions}{All known corresponding accession IDs in INSDC (GenBank/ENA/DDBJ) ideally with a version number, for example NC_000001}

\item{sourceURI}{The URI from which the references were obtained}
}
\value{
ReferenceSet object
}
\description{
ReferenceSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference set is a set of references which typically comprise a reference assembly for a species, such as GRCh38 which is representative of the human genome. A reference set defines a common coordinate space for comparing reference-aligned experimental data. A reference set contains 1 or more references.
}

