% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{referencesets.get}
\alias{referencesets.get}
\title{Gets a reference set.Implements GlobalAllianceApi.getReferenceSet.}
\usage{
referencesets.get(referenceSetId)
}
\arguments{
\item{referenceSetId}{The ID of the reference set}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
\item https://www.googleapis.com/auth/genomics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics, https://www.googleapis.com/auth/genomics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}
}

