% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchAnnotationSetsRequest}
\alias{SearchAnnotationSetsRequest}
\title{SearchAnnotationSetsRequest Object}
\usage{
SearchAnnotationSetsRequest(datasetIds = NULL, name = NULL,
  pageSize = NULL, pageToken = NULL, referenceSetId = NULL,
  types = NULL)
}
\arguments{
\item{datasetIds}{The dataset IDs to search within}

\item{name}{Only return annotations sets for which a substring of the name matches this string (case insensitive)}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{referenceSetId}{If specified, only annotation sets associated with the given reference set are returned}

\item{types}{If specified, only annotation sets that have any of these types are returned}
}
\value{
SearchAnnotationSetsRequest object
}
\description{
SearchAnnotationSetsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SearchAnnotationSetsRequest functions: \code{\link{annotationSets.search}}
}

