% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchAnnotationsRequest}
\alias{SearchAnnotationsRequest}
\title{SearchAnnotationsRequest Object}
\usage{
SearchAnnotationsRequest(annotationSetIds = NULL, pageSize = NULL,
  pageToken = NULL, range = NULL)
}
\arguments{
\item{annotationSetIds}{The annotation sets to search within}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{range}{If specified, this query matches only annotations that overlap this range}
}
\value{
SearchAnnotationsRequest object
}
\description{
SearchAnnotationsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SearchAnnotationsRequest functions: \code{\link{annotations.search}}
}

