% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchCallSetsRequest}
\alias{SearchCallSetsRequest}
\title{SearchCallSetsRequest Object}
\usage{
SearchCallSetsRequest(name = NULL, pageSize = NULL, pageToken = NULL,
  variantSetIds = NULL)
}
\arguments{
\item{name}{Only return call sets for which a substring of the name matches this string}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{variantSetIds}{Restrict the query to call sets within the given variant sets}
}
\value{
SearchCallSetsRequest object
}
\description{
SearchCallSetsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The call set search request.
}
\seealso{
Other SearchCallSetsRequest functions: \code{\link{callsets.search}}
}

