% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchJobsRequest}
\alias{SearchJobsRequest}
\title{SearchJobsRequest Object}
\usage{
SearchJobsRequest(createdAfter = NULL, createdBefore = NULL,
  pageSize = NULL, pageToken = NULL, projectNumber = NULL,
  status = NULL)
}
\arguments{
\item{createdAfter}{If specified, only jobs created on or after this date, given in milliseconds since Unix epoch, will be returned}

\item{createdBefore}{If specified, only jobs created prior to this date, given in milliseconds since Unix epoch, will be returned}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token which is used to page through large result sets}

\item{projectNumber}{Required}

\item{status}{Only return jobs which have a matching status}
}
\value{
SearchJobsRequest object
}
\description{
SearchJobsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The jobs search request.
}
\seealso{
Other SearchJobsRequest functions: \code{\link{jobs.search}}
}

