% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchJobsResponse}
\alias{SearchJobsResponse}
\title{SearchJobsResponse Object}
\usage{
SearchJobsResponse(jobs = NULL, nextPageToken = NULL)
}
\arguments{
\item{jobs}{The list of jobs results, ordered newest to oldest}

\item{nextPageToken}{The continuation token which is used to page through large result sets}
}
\value{
SearchJobsResponse object
}
\description{
SearchJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The job search response.
}

