% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReadGroupSetsRequest}
\alias{SearchReadGroupSetsRequest}
\title{SearchReadGroupSetsRequest Object}
\usage{
SearchReadGroupSetsRequest(datasetIds = NULL, name = NULL,
  pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{datasetIds}{Restricts this query to read group sets within the given datasets}

\item{name}{Only return read group sets for which a substring of the name matches this string}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}
}
\value{
SearchReadGroupSetsRequest object
}
\description{
SearchReadGroupSetsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read group set search request.
}
\seealso{
Other SearchReadGroupSetsRequest functions: \code{\link{readgroupsets.search}}
}

