% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReadsRequest}
\alias{SearchReadsRequest}
\title{SearchReadsRequest Object}
\usage{
SearchReadsRequest(end = NULL, pageSize = NULL, pageToken = NULL,
  readGroupIds = NULL, readGroupSetIds = NULL, referenceName = NULL,
  start = NULL)
}
\arguments{
\item{end}{The end position of the range on the reference, 0-based exclusive}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{readGroupIds}{The IDs of the read groups within which to search for reads}

\item{readGroupSetIds}{The IDs of the read groups sets within which to search for reads}

\item{referenceName}{The reference sequence name, for example chr1, 1, or chrX}

\item{start}{The start position of the range on the reference, 0-based inclusive}
}
\value{
SearchReadsRequest object
}
\description{
SearchReadsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read search request.
}
\seealso{
Other SearchReadsRequest functions: \code{\link{reads.search}}
}

