% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReadsResponse}
\alias{SearchReadsResponse}
\title{SearchReadsResponse Object}
\usage{
SearchReadsResponse(alignments = NULL, nextPageToken = NULL)
}
\arguments{
\item{alignments}{The list of matching alignments sorted by mapped genomic coordinate, if any, ascending in position within the same reference}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}
}
\value{
SearchReadsResponse object
}
\description{
SearchReadsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read search response.
}

