% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReferenceSetsRequest}
\alias{SearchReferenceSetsRequest}
\title{SearchReferenceSetsRequest Object}
\usage{
SearchReferenceSetsRequest(accessions = NULL, assemblyId = NULL,
  md5checksums = NULL, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{accessions}{If present, return references for which the accession matches any of these strings}

\item{assemblyId}{If present, return reference sets for which a substring of their assemblyId matches this string (case insensitive)}

\item{md5checksums}{If present, return references for which the md5checksum matches}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}
}
\value{
SearchReferenceSetsRequest object
}
\description{
SearchReferenceSetsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SearchReferenceSetsRequest functions: \code{\link{referencesets.search}}
}

