% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchReferencesRequest}
\alias{SearchReferencesRequest}
\title{SearchReferencesRequest Object}
\usage{
SearchReferencesRequest(accessions = NULL, md5checksums = NULL,
  pageSize = NULL, pageToken = NULL, referenceSetId = NULL)
}
\arguments{
\item{accessions}{If present, return references for which the accession matches this string}

\item{md5checksums}{If present, return references for which the md5checksum matches}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{referenceSetId}{If present, return only references which belong to this reference set}
}
\value{
SearchReferencesRequest object
}
\description{
SearchReferencesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SearchReferencesRequest functions: \code{\link{references.search}}
}

