% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchVariantSetsRequest}
\alias{SearchVariantSetsRequest}
\title{SearchVariantSetsRequest Object}
\usage{
SearchVariantSetsRequest(datasetIds = NULL, pageSize = NULL,
  pageToken = NULL)
}
\arguments{
\item{datasetIds}{Exactly one dataset ID must be provided here}

\item{pageSize}{The maximum number of results to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}
}
\value{
SearchVariantSetsRequest object
}
\description{
SearchVariantSetsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The search variant sets request.
}
\seealso{
Other SearchVariantSetsRequest functions: \code{\link{variantsets.search}}
}

