% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchVariantsRequest}
\alias{SearchVariantsRequest}
\title{SearchVariantsRequest Object}
\usage{
SearchVariantsRequest(callSetIds = NULL, end = NULL, maxCalls = NULL,
  pageSize = NULL, pageToken = NULL, referenceName = NULL, start = NULL,
  variantName = NULL, variantSetIds = NULL)
}
\arguments{
\item{callSetIds}{Only return variant calls which belong to call sets with these ids}

\item{end}{The end of the window, 0-based exclusive}

\item{maxCalls}{The maximum number of calls to return in a single page}

\item{pageSize}{The maximum number of variants to return in a single page}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{referenceName}{Required}

\item{start}{The beginning of the window (0-based, inclusive) for which overlapping variants should be returned}

\item{variantName}{Only return variants which have exactly this name}

\item{variantSetIds}{At most one variant set ID must be provided}
}
\value{
SearchVariantsRequest object
}
\description{
SearchVariantsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant search request.
}
\seealso{
Other SearchVariantsRequest functions: \code{\link{variants.search}}
}

