% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{SearchVariantsResponse}
\alias{SearchVariantsResponse}
\title{SearchVariantsResponse Object}
\usage{
SearchVariantsResponse(nextPageToken = NULL, variants = NULL)
}
\arguments{
\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{variants}{The list of matching Variants}
}
\value{
SearchVariantsResponse object
}
\description{
SearchVariantsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The variant search response.
}

