% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Transcript}
\alias{Transcript}
\title{Transcript Object}
\usage{
Transcript(codingSequence = NULL, exons = NULL, geneId = NULL)
}
\arguments{
\item{codingSequence}{The range of the coding sequence for this transcript, if any}

\item{exons}{The exons that compose this transcript}

\item{geneId}{The annotation ID of the gene from which this transcript is transcribed}
}
\value{
Transcript object
}
\description{
Transcript Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A transcript represents the assertion that a particular region of the reference genome may be transcribed as RNA.
}

