% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{TranscriptCodingSequence}
\alias{TranscriptCodingSequence}
\title{TranscriptCodingSequence Object}
\usage{
TranscriptCodingSequence(end = NULL, start = NULL)
}
\arguments{
\item{end}{The end of the coding sequence on this annotation's reference sequence, 0-based exclusive}

\item{start}{The start of the coding sequence on this annotation's reference sequence, 0-based inclusive}
}
\value{
TranscriptCodingSequence object
}
\description{
TranscriptCodingSequence Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

