% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{TranscriptExon}
\alias{TranscriptExon}
\title{TranscriptExon Object}
\usage{
TranscriptExon(end = NULL, frame = NULL, start = NULL)
}
\arguments{
\item{end}{The end position of the exon on this annotation's reference sequence, 0-based exclusive}

\item{frame}{The frame of this exon}

\item{start}{The start position of the exon on this annotation's reference sequence, 0-based inclusive}
}
\value{
TranscriptExon object
}
\description{
TranscriptExon Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

