% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{Variant.info}
\alias{Variant.info}
\title{Variant.info Object}
\usage{
Variant.info()
}
\value{
Variant.info object
}
\description{
Variant.info Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A string which maps to an array of values.
}
\seealso{
Other Variant functions: \code{\link{Variant}},
  \code{\link{variants.create}},
  \code{\link{variants.update}}
}

