% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantAnnotation}
\alias{VariantAnnotation}
\title{VariantAnnotation Object}
\usage{
VariantAnnotation(alternateBases = NULL, clinicalSignificance = NULL,
  conditions = NULL, effect = NULL, geneId = NULL, transcriptIds = NULL,
  type = NULL)
}
\arguments{
\item{alternateBases}{The alternate allele for this variant}

\item{clinicalSignificance}{Describes the clinical significance of a variant}

\item{conditions}{The set of conditions associated with this variant}

\item{effect}{Effect of the variant on the coding sequence}

\item{geneId}{Google annotation ID of the gene affected by this variant}

\item{transcriptIds}{Google annotation IDs of the transcripts affected by this variant}

\item{type}{Type has been adapted from ClinVar's list of variant types}
}
\value{
VariantAnnotation object
}
\description{
VariantAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Variant annotation.
}

