% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantAnnotationCondition}
\alias{VariantAnnotationCondition}
\title{VariantAnnotationCondition Object}
\usage{
VariantAnnotationCondition(conceptId = NULL, externalIds = NULL,
  names = NULL, omimId = NULL)
}
\arguments{
\item{conceptId}{The MedGen concept id associated with this gene}

\item{externalIds}{The set of external IDs for this condition}

\item{names}{A set of names for the condition}

\item{omimId}{The OMIM id for this condition}
}
\value{
VariantAnnotationCondition object
}
\description{
VariantAnnotationCondition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

