% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variants.update}
\alias{variants.update}
\title{Updates a variant's names and info fields. All other modifications are silently ignored. Returns the modified variant without its calls.}
\usage{
variants.update(Variant, variantId)
}
\arguments{
\item{Variant}{The \link{Variant} object to pass to this method}

\item{variantId}{The ID of the variant to be updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}

Other Variant functions: \code{\link{Variant.info}},
  \code{\link{Variant}}, \code{\link{variants.create}}
}

