% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_objects.R
\name{VariantSet}
\alias{VariantSet}
\title{VariantSet Object}
\usage{
VariantSet(datasetId = NULL, description = NULL, id = NULL,
  metadata = NULL, name = NULL, referenceBounds = NULL,
  referenceSetId = NULL)
}
\arguments{
\item{datasetId}{The dataset to which this variant set belongs}

\item{description}{A textual description of this variant set}

\item{id}{The Google-generated ID of the variant set}

\item{metadata}{The metadata associated with this variant set}

\item{name}{User-specified, mutable name}

\item{referenceBounds}{A list of all references used by the variants in a variant set with associated coordinate upper bounds for each one}

\item{referenceSetId}{The reference set to which the variant set is mapped}
}
\value{
VariantSet object
}
\description{
VariantSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A variant set is a collection of call sets and variants. It contains summary statistics of those contents. A variant set belongs to a dataset.
}
\seealso{
Other VariantSet functions: \code{\link{variantsets.create}},
  \code{\link{variantsets.patch}},
  \code{\link{variantsets.update}}
}

