% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variantsets.export}
\alias{variantsets.export}
\title{Exports variant set data to an external destination.}
\usage{
variantsets.export(ExportVariantSetRequest, variantSetId)
}
\arguments{
\item{ExportVariantSetRequest}{The \link{ExportVariantSetRequest} object to pass to this method}

\item{variantSetId}{The ID of the variant set that contains variant data which should be exported}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}

Other ExportVariantSetRequest functions: \code{\link{ExportVariantSetRequest}}
}

