% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics_functions.R
\name{variantsets.mergeVariants}
\alias{variantsets.mergeVariants}
\title{Merges the given variants with existing variants. Each variant will be merged with an existing variant that matches its reference sequence, start, end, reference bases, and alternative bases. If no such variant exists, a new one will be created.When variants are merged, the call information from the new variant is added to the existing variant. Variant info fields are merged as specified in the InfoMergeConfig field of the MergeVariantsRequest.}
\usage{
variantsets.mergeVariants(MergeVariantsRequest, variantSetId)
}
\arguments{
\item{MergeVariantsRequest}{The \link{MergeVariantsRequest} object to pass to this method}

\item{variantSetId}{The destination variant set}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/genomics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/genomics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/genomics/v1beta2/reference}{Google Documentation}

Other MergeVariantsRequest functions: \code{\link{MergeVariantsRequest}}
}

